backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
python-docx>=0.8.11
PyPDF2>=3.0.1
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
nltk>=3.8.1
textstat>=0.7.3
python-dotenv>=1.0.0
requests>=2.31.0
